move net_op to operators
